Update forest-cover-loss-estimation.js
Thanks so much Justin for all the comments! I learned a lot in the process. It is indeed quite tricky when it comes to scaling and reducing the area. I also used 'selfMask()' extensively throughout. I hope that users place the selfMask appropriately, as misplacing them can lead to wrong figures and maps.