Add histogram charts of slope for each national park
Addressing @jdbcode Comment: Add histogram charts of slope for each region to more easily convey the distribution of slope values. The map display sort of does this, but the histogram is more quantitative. It would be nice to add the two histograms as a side-by-side image in the .md file as a .png figure.