Create GOES-16-17_FireReclassification
This script uses the 'Mask' band of the GOES-16 and GOES-17 Fire Detection and Characterization (FDC) product to identify pixels with active fires. The 'Mask' band is more informative than the 'DQF' band. Pixels with fire mask classes 10-14 and 30-34 have the highest likelihoods of active fire. These identified fire pixels are buffered to stand out. Every fire mask class, from 0 to 245, is also reclassified into one of twelve descriptive categories which indicate the likelihood of fire, the quality of data, and a rationale if there is poor data quality. Additionally, pixels which have data on the fire's radiative power (FRP) -- a metric of the fire intensity-- are displayed with a yellow-orange-red color ramp in another map layer.